+++ CHANGES.rst
@@ -1,24 +1,47 @@
 Changelog
 =========
 
-1.2 (unreleased)
-----------------
 
-- incompatible changes (none yet)
-  [tobiasherp]
-
-
-1.1.2 (unreleased)
+1.1.4 (2019-05-09)
 ------------------
 
-- interfaces module
+- New features:
+
+  - Imported some more permissions from ``AccessControl.Permissions``
+    and ``Products.CMFCore.permissions``
+
+[tobiasherp]
+
+
+1.1.3 (2019-01-31)
+------------------
+
+- New features:
+
+  - New permission string constants for ``visaplan.plone.browsers``
+
+[tobiasherp]
+
+
+1.1.2 (2018-11-22)
+------------------
+
+- Bugfix for getContext: ``unbound`` module
+  [tobiasherp]
+
+- New permission string constants for ``visaplan.plone.pdfexports``
+  [tobiasherp]
+
+- ``interfaces`` module
+
+- Package can be (partly) used even without Zope/Plone
   [tobiasherp]
 
 
 1.1.1 (2018-09-26)
 ------------------
 
-- New exception MissingFileError
+- New exception ``MissingFileError``
   [tobiasherp]